moved reports related interfaces to reportsInterfaces, moved planning related interfaces to planningInterfaces
