+++ src/components/Report/ReportRow.tsx
@@ -1,7 +1,7 @@
-import { Button } from 'hds-react';
-import { FC, useCallback } from 'react';
+import { Button, IconDownload } from 'hds-react';
+import { FC, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ReportType } from '@/interfaces/common';
+import { ReportType } from '@/interfaces/reportInterfaces';
 import './styles.css';
 
 interface IReportRowProps {
@@ -11,6 +11,8 @@ interface IReportRowProps {
 
 const ReportRow: FC<IReportRowProps> = ({ type, lastUpdated }) => {
   const { t } = useTranslation();
+
+  const downloadIcon = useMemo(() => <IconDownload />, []);
 
   const buildPdf = useCallback(() => {
     // TODO: handle building pdf for different types
@@ -30,19 +32,23 @@ const ReportRow: FC<IReportRowProps> = ({ type, lastUpdated }) => {
 
   return (
     <div className="report-row-container" data-testid={`report-row-${type}`}>
+      {/* report title */}
       <h3 className="report-title" data-testid={`report-title-${type}`}>
         {t(`reports.${type}.title`)}
       </h3>
+      {/* last updated date */}
       <div className="report-last-updated" data-testid={`last-updated-${type}`}>{`${t(
         'lastUpdated',
-      )}: ${lastUpdated}`}</div>
+      )} ${lastUpdated}`}</div>
+      {/* download pdf button */}
       <div className="report-download-pdf-button" data-testid={`download-pdf-${type}`}>
-        <Button onClick={buildPdf} disabled={true}>
+        <Button iconLeft={downloadIcon} onClick={buildPdf} disabled={true}>
           {t('downloadPdf', { name: t(`reports.${type}.documentName`) })}
         </Button>
       </div>
+      {/* download xlsx button */}
       <div className="report-download-xlsx-button" data-testid={`download-xlsx-${type}`}>
-        <Button variant="secondary" onClick={buildXlsx} disabled={true}>
+        <Button iconLeft={downloadIcon} variant="secondary" onClick={buildXlsx} disabled={true}>
           {t('downloadXlsx', { name: t(`reports.${type}.documentName`) })}
         </Button>
       </div>
