+++ src/interfaces/reportInterfaces.ts
@@ -1,0 +1,9 @@
+export const reports = [
+  'budgetProposal',
+  'strategy',
+  'constructionProgram',
+  'budgetBookSummary',
+  'financialStatement',
+] as const;
+
+export type ReportType = (typeof reports)[number];
